🚀 Permissive v0.0.4

+++ src/core/PermissiveAccount.sol
@@ -30,10 +30,7 @@ contract PermissiveAccount is BaseAccount, IPermissiveAccount, Ownable, EIP712 {
     FeeManager private immutable feeManager;
     bool private _initialized;
 
-    constructor(
-        address __entryPoint,
-        address payable _feeManager
-    ) EIP712("Permissive Account", "v0.0.4") {
+    constructor(address __entryPoint, address payable _feeManager) EIP712("Permissive Account", "v0.0.4") {
         _entryPoint = IEntryPoint(__entryPoint);
         feeManager = FeeManager(_feeManager);
     }
@@ -52,35 +49,17 @@ contract PermissiveAccount is BaseAccount, IPermissiveAccount, Ownable, EIP712 {
         _transferOwnership(owner);
     }
 
-    function setOperatorPermissions(
-        PermissionSet calldata permSet,
-        bytes calldata signature
-    ) external {
-        bytes32 digest = _hashTypedDataV4(
-            keccak256(
-                abi.encode(
-                    typedStruct,
-                    permSet.operator,
-                    permSet.merkleRootPermissions
-                )
-            )
-        );
+    function setOperatorPermissions(PermissionSet calldata permSet, bytes calldata signature) external {
+        bytes32 digest =
+            _hashTypedDataV4(keccak256(abi.encode(typedStruct, permSet.operator, permSet.merkleRootPermissions)));
         address signer = ECDSA.recover(digest, signature);
         if (signer != owner()) revert("Not Allowed");
         bytes32 oldValue = operatorPermissions[permSet.operator];
         operatorPermissions[permSet.operator] = permSet.merkleRootPermissions;
-        emit OperatorMutated(
-            permSet.operator,
-            oldValue,
-            permSet.merkleRootPermissions
-        );
+        emit OperatorMutated(permSet.operator, oldValue, permSet.merkleRootPermissions);
     }
 
-    function validateUserOp(
-        UserOperation calldata userOp,
-        bytes32 userOpHash,
-        uint256 missingAccountFunds
-    )
+    function validateUserOp(UserOperation calldata userOp, bytes32 userOpHash, uint256 missingAccountFunds)
         external
         override(BaseAccount, IAccount)
         returns (uint256 validationData)
@@ -88,59 +67,26 @@ contract PermissiveAccount is BaseAccount, IPermissiveAccount, Ownable, EIP712 {
         _requireFromEntryPoint();
         bytes32 hash = userOpHash.toEthSignedMessageHash();
         if (owner() != hash.recover(userOp.signature)) {
-            (
-                ,
-                ,
-                ,
-                PermissionLib.Permission memory permission,
-                bytes32[] memory proof,
-                uint256 providedFee
-            ) = abi.decode(
-                    userOp.callData[4:],
-                    (
-                        address,
-                        uint256,
-                        bytes,
-                        PermissionLib.Permission,
-                        bytes32[],
-                        uint256
-                    )
-                );
+            (,,, PermissionLib.Permission memory permission, bytes32[] memory proof, uint256 providedFee) =
+                abi.decode(userOp.callData[4:], (address, uint256, bytes, PermissionLib.Permission, bytes32[], uint256));
             if (permission.operator.code.length > 0) {
-                try
-                    IERC1271(permission.operator).isValidSignature(
-                        hash,
-                        userOp.signature
-                    )
-                returns (bytes4 magicValue) {
+                try IERC1271(permission.operator).isValidSignature(hash, userOp.signature) returns (bytes4 magicValue) {
                     validationData = _packValidationData(
                         ValidationData(
-                            magicValue == IERC1271.isValidSignature.selector
-                                ? address(0)
-                                : address(1),
+                            magicValue == IERC1271.isValidSignature.selector ? address(0) : address(1),
                             permission.validAfter,
                             permission.validUntil
                         )
                     );
                 } catch {
-                    validationData = _packValidationData(
-                        ValidationData(
-                            address(1),
-                            permission.validAfter,
-                            permission.validUntil
-                        )
-                    );
+                    validationData =
+                        _packValidationData(ValidationData(address(1), permission.validAfter, permission.validUntil));
                 }
             } else if (permission.operator != hash.recover(userOp.signature)) {
                 return SIG_VALIDATION_FAILED;
             } else {
-                validationData = _packValidationData(
-                    ValidationData(
-                        address(0),
-                        permission.validAfter,
-                        permission.validUntil
-                    )
-                );
+                validationData =
+                    _packValidationData(ValidationData(address(0), permission.validAfter, permission.validUntil));
             }
             bytes32 permHash = permission.hash();
             _validateMerklePermission(permission, proof, permHash);
@@ -163,23 +109,11 @@ contract PermissiveAccount is BaseAccount, IPermissiveAccount, Ownable, EIP712 {
         uint256 gasFee
     ) external {
         _requireFromEntryPointOrOwner();
-        payable(address(feeManager)).transfer(
-            (gasFee * feeManager.fee()) / 10000
-        );
+        payable(address(feeManager)).transfer((gasFee * feeManager.fee()) / 10000);
         (bool success, bytes memory result) = dest.call{value: value}(
-            bytes.concat(
-                func.slice(0, 4),
-                AllowanceCalldata.RLPtoABI(func.slice(4, func.length - 4))
-            )
+            bytes.concat(func.slice(0, 4), AllowanceCalldata.RLPtoABI(func.slice(4, func.length - 4)))
         );
-        emit PermissionUsed(
-            permission.hash(),
-            dest,
-            value,
-            func,
-            permission,
-            gasFee
-        );
+        emit PermissionUsed(permission.hash(), dest, value, func, permission, gasFee);
         if (!success) {
             assembly {
                 revert(add(result, 32), mload(result))
@@ -187,18 +121,10 @@ contract PermissiveAccount is BaseAccount, IPermissiveAccount, Ownable, EIP712 {
         }
     }
 
-    function computeGasFee(
-        UserOperation memory userOp
-    ) public pure returns (uint256 fee) {
+    function computeGasFee(UserOperation memory userOp) public pure returns (uint256 fee) {
         unchecked {
-            uint256 mul = address(bytes20(userOp.paymasterAndData)) !=
-                address(0)
-                ? 3
-                : 1;
-            uint256 requiredGas = userOp.callGasLimit +
-                userOp.verificationGasLimit *
-                mul +
-                userOp.preVerificationGas;
+            uint256 mul = address(bytes20(userOp.paymasterAndData)) != address(0) ? 3 : 1;
+            uint256 requiredGas = userOp.callGasLimit + userOp.verificationGasLimit * mul + userOp.preVerificationGas;
 
             fee = requiredGas * userOp.maxFeePerGas;
         }
@@ -206,21 +132,16 @@ contract PermissiveAccount is BaseAccount, IPermissiveAccount, Ownable, EIP712 {
 
     /* INTERNAL */
 
-    function _hashTypedDataV4(
-        bytes32 structHash
-    ) internal view override returns (bytes32) {
+    function _hashTypedDataV4(bytes32 structHash) internal view override returns (bytes32) {
         return ECDSA.toTypedDataHash(_domainSeparatorV4(), structHash);
     }
 
-    function _validateData(
-        PermissionLib.Permission memory permission
-    ) internal view {
+    function _validateData(PermissionLib.Permission memory permission) internal view {
         if (
-            permission.dataValidation.validator != address(0) &&
-            !IDataValidator(permission.dataValidation.validator).isValidData(
-                permission.dataValidation.target,
-                permission.dataValidation.data
-            )
+            permission.dataValidation.validator != address(0)
+                && !IDataValidator(permission.dataValidation.validator).isValidData(
+                    permission.dataValidation.target, permission.dataValidation.data
+                )
         ) {
             revert("Invalid data");
         }
@@ -231,12 +152,10 @@ contract PermissiveAccount is BaseAccount, IPermissiveAccount, Ownable, EIP712 {
         PermissionLib.Permission memory permission,
         bytes32 permHash
     ) internal {
-        (address to, uint256 value, bytes memory callData, , ) = abi.decode(
-            userOp.callData[4:],
-            (address, uint256, bytes, PermissionLib.Permission, bytes32[])
-        );
+        (address to, uint256 value, bytes memory callData,,) =
+            abi.decode(userOp.callData[4:], (address, uint256, bytes, PermissionLib.Permission, bytes32[]));
         if (permission.to != to) revert("InvalidTo");
-        uint rPermU = remainingPermUsage[permHash];
+        uint256 rPermU = remainingPermUsage[permHash];
         if (permission.maxUsage > 0) {
             if (permission.maxUsage == 1) revert("OutOfPerms");
             if (rPermU == 1) {
@@ -250,9 +169,7 @@ contract PermissiveAccount is BaseAccount, IPermissiveAccount, Ownable, EIP712 {
         }
         if (
             !AllowanceCalldata.isAllowedCalldata(
-                permission.allowed_arguments,
-                callData.slice(4, callData.length - 4),
-                value
+                permission.allowed_arguments, callData.slice(4, callData.length - 4), value
             )
         ) revert("Not allowed Calldata");
         if (permission.selector != bytes4(callData)) revert("InvalidSelector");
@@ -271,25 +188,21 @@ contract PermissiveAccount is BaseAccount, IPermissiveAccount, Ownable, EIP712 {
         bytes32[] memory proof,
         bytes32 permHash
     ) internal view {
-        bool isValidProof = MerkleProof.verify(
-            proof,
-            operatorPermissions[permission.operator],
-            keccak256(bytes.concat(permHash))
-        );
+        bool isValidProof =
+            MerkleProof.verify(proof, operatorPermissions[permission.operator], keccak256(bytes.concat(permHash)));
         if (!isValidProof) revert("Invalid Proof");
     }
 
     function _requireFromEntryPointOrOwner() internal view {
-        require(
-            msg.sender == address(entryPoint()) || msg.sender == owner(),
-            "account: not from EntryPoint or owner"
-        );
+        require(msg.sender == address(entryPoint()) || msg.sender == owner(), "account: not from EntryPoint or owner");
     }
 
-    function _validateSignature(
-        UserOperation calldata userOp,
-        bytes32 userOpHash
-    ) internal view override returns (uint256 validationData) {
+    function _validateSignature(UserOperation calldata userOp, bytes32 userOpHash)
+        internal
+        view
+        override
+        returns (uint256 validationData)
+    {
         bytes32 hash = userOpHash.toEthSignedMessageHash();
         if (owner() != hash.recover(userOp.signature)) {
             return SIG_VALIDATION_FAILED;
@@ -299,18 +212,12 @@ contract PermissiveAccount is BaseAccount, IPermissiveAccount, Ownable, EIP712 {
 
     function _payPrefund(uint256 missingAccountFunds) internal override {
         if (missingAccountFunds != 0) {
-            (bool success, ) = payable(msg.sender).call{
-                value: missingAccountFunds,
-                gas: type(uint256).max
-            }("");
+            (bool success,) = payable(msg.sender).call{value: missingAccountFunds, gas: type(uint256).max}("");
             (success);
         }
     }
 
-    function isValidSignature(
-        bytes32 _hash,
-        bytes calldata _signature
-    ) external view returns (bytes4) {
+    function isValidSignature(bytes32 _hash, bytes calldata _signature) external view returns (bytes4) {
         if (ECDSA.recover(_hash, _signature) == owner()) {
             return 0x1626ba7e;
         } else {
